requirements.txt
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
django-debug-toolbar==6.3.0
django-filter==25.2
django-graphiql-debug-toolbar==0.2.0
django-htmx==1.27.0
django-mptt==0.18.0
django-pglocks==1.0.4
django-prometheus==2.4.0
django-redis==6.0.0
django-rich==2.2.0
django-rq==4.0.1
django-storages==1.14.6
django-tables2==2.8.0
django-taggit==6.1.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.4.1
feedparser==6.0.12
gunicorn==25.3.0
Jinja2==3.1.6
jsonschema==4.26.0
Markdown==3.10.2
mkdocs==1.6.1
mkdocs-material==9.7.6
mkdocstrings==1.0.3
mkdocstrings-python==2.0.3
netaddr==1.3.0
nh3==0.3.4
Pillow==12.2.0
psycopg[c,pool]==3.3.3
PyYAML==6.0.3
requests==2.33.1
rq==2.7.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sorl-thumbnail==13.0.0
strawberry-graphql==0.312.2
strawberry-graphql-django==0.82.1
svgwrite==1.4.3
tablib==3.9.0
tzdata==2026.1
zensical==0.0.28
